Make sure pages are mprotected PROT_EXEC for amd64 support

+++ gcl/o/gbc.c
@@ -125,7 +125,7 @@ struct apage {
   char apage_self[PAGESIZE];
 };
 
-int maxpage;
+long maxpage;
 
 object sSAnotify_gbcA;
